Fix udp send for clients on lan
When sending udp packets on lan, it is necessary to have the destination mac. If the device is just a client, it can issue a broadcast message to find the server and when it gets a reply, stores the mac address of the server that will be automatically used when sending udp packets to this server.